refactor: move code for hash-object command in object package
